W-11757246 0/3 (refactor): made payload validation tests use scala interfaces instead of platform
