+++ amf-cli/js/src/test/scala/amf/client/validation/JsClientPayloadValidationTest.scala
@@ -1,0 +1,5 @@
+package amf.client.validation
+
+import amf.cli.internal.convert.NativeOpsFromJs
+
+class JsClientPayloadValidationTest extends ClientPayloadValidationTest with NativeOpsFromJs {}
