+++ amf-cli/js/src/test/scala/amf/client/validation/JsPayloadValidationTest.scala
@@ -2,4 +2,4 @@ package amf.client.validation
 
 import amf.cli.internal.convert.NativeOpsFromJs
 
-class JsPayloadValidationTest extends ClientPayloadValidationTest with NativeOpsFromJs
+class JsPayloadValidationTest extends PayloadValidationTest with NativeOpsFromJs
