+++ amf-cli/jvm/src/test/scala/amf/client/validation/JvmClientPayloadValidationTest.scala
@@ -1,0 +1,5 @@
+package amf.client.validation
+
+import amf.cli.internal.convert.NativeOpsFromJvm
+
+class JvmClientPayloadValidationTest extends ClientPayloadValidationTest with NativeOpsFromJvm
